requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.9.2
click==8.1.3
email-validator==2.0.0.post2
flask==2.2.2
flask-cors==3.0.10
flask-login==0.6.2
flask-migrate==4.0.2
flask-sqlalchemy==3.0.2
flask-wtf==1.1.1
googlemaps==4.2.0
greenlet==3.1.1
gunicorn==20.1.0
itsdangerous==2.1.2
jinja2==3.1.2
mako==1.2.4
markupsafe==2.1.2
openai==0.27.8
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
six==1.16.0
sqlalchemy==1.4.46
werkzeug==2.2.2
wtforms==3.0.1
boto3==1.35.83
